+++ libconcurrent/includes/ccqueue.h
@@ -7,8 +7,8 @@
 /// Panagiota Fatourou, and Nikolaos D. Kallimanis."Revisiting the combining synchronization technique".
 /// ACM SIGPLAN Notices. Vol. 47. No. 8. ACM, PPoPP 2012.
 /// @copyright Copyright (c) 2021
-#ifndef _CCSTACK_H_
-#define _CCSTACK_H_
+#ifndef _CCQUEUE_H_
+#define _CCQUEUE_H_
 
 #include <config.h>
 #include <queue-stack.h>